// web/lib/api/generated/model/priceChange.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface PriceChange {
  code: string;
  name: string;
  /**
     * @nullable
     * @pattern ^-?\d{0,12}(?:\.\d{0,2})?$
     */
  old: string | null;
  /** @pattern ^-?\d{0,12}(?:\.\d{0,2})?$ */
  new: string;
}